requirements.txt
google-genai
python-dotenv
requests
gradio
loguru